Refactor: Move `ServiceContainer` and `ServiceInterface` to `Engine\Service` namespace and update references
