+++ Engine/Service/ServiceInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Engine;
+namespace Engine\Service;
 
 /**
  * Interface ServiceInterface
