+++ application/modules/Health/Module.php
@@ -5,7 +5,7 @@ namespace Modules\Health;
 use Engine\Modules\AbstractModule;
 use Engine\Router;
 use Engine\Routing\AttributeRouteLoader;
-use Engine\ServiceContainer;
+use Engine\Service\ServiceContainer;
 use Modules\Health\Controllers\HealthController;
 use Modules\Health\Services\HealthService;
